@@ -11,7 +11,7 @@ class AppDatabase {
   static final AppDatabase _singleton = AppDatabase._();
 
   static AppDatabase get instance => _singleton;
-  Completer<Database> _dbOpenCompleter;
+  Completer<Database>? _dbOpenCompleter;
 
   AppDatabase._();
 
@@ -20,7 +20,7 @@ class AppDatabase {
       _dbOpenCompleter = Completer();
       _openDatabase();
     }
-    return _dbOpenCompleter.future;
+    return _dbOpenCompleter!.future;
   }
 
   Future _openDatabase() async {
@@ -29,6 +29,6 @@ class AppDatabase {
     print(dbPath);
 
     final database = await databaseFactoryIo.openDatabase(dbPath);
-    _dbOpenCompleter.complete(database);
+    _dbOpenCompleter!.complete(database);
   }
 }
