@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:sembast/sembast.dart';
 import 'package:what_when/model/app_database.dart';
 import 'package:what_when/model/serializers.dart';
@@ -5,7 +7,8 @@ import 'package:what_when/model/TaskModel.dart';
 
 class TaskDAO {
   static const String taskStoreName = 'tasks';
-  final _taskStore = intMapStoreFactory.store(taskStoreName);
+  final StoreRef<int?, Map<String, Object?>> _taskStore =
+      intMapStoreFactory.store(taskStoreName);
 
   static final TaskDAO _singleton = TaskDAO._();
 
@@ -16,16 +19,19 @@ class TaskDAO {
 
   //insert into database method
   Future insert(TaskModel task) async {
-    await _taskStore.record(task.id).put(await _db,
-        standardSerializers.serializeWith(TaskModel.serializer, task));
+    await _taskStore.record(task.id).put(
+        await _db,
+        standardSerializers.serializeWith(TaskModel.serializer, task)
+            as Map<String, Object?>);
   }
 
   Future getAll() async {
     return await _taskStore.find(await _db);
   }
 
-  Future<TaskModel> getTaskById(int id) async {
-    var task = await _taskStore.record(id).get(await _db);
+  Future<TaskModel?> getTaskById(int id) async {
+    var task = await (_taskStore.record(id).get(await _db)
+        as FutureOr<Map<String, Object>?>);
     return standardSerializers.deserializeWith(TaskModel.serializer, task);
   }
 
