@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:collection';
 
 import 'package:flutter/material.dart';
@@ -9,35 +10,36 @@ import 'package:what_when/model/TaskModel.dart';
 
 class TaskListModel extends ChangeNotifier {
   static const String TASK_STORE_NAME = 'tasks';
-  final _taskStore = intMapStoreFactory.store(TASK_STORE_NAME);
+  final StoreRef<int?, Map<String, Object?>> _taskStore =
+      intMapStoreFactory.store(TASK_STORE_NAME);
 
   // Private getter to shorten the amount of code needed to get the
   // singleton instance of an opened database.
 
   Future<Database> get _db async => await AppDatabase.instance.database;
 
-  List<TaskModel> _tasks = [];
+  List<TaskModel?> _tasks = [];
 
   int get getListLength => _tasks.length;
 
-  UnmodifiableListView<TaskModel> get getTaskList =>
+  UnmodifiableListView<TaskModel?> get getTaskList =>
       UnmodifiableListView(_tasks);
 
   get getLastId {
-    return _tasks.isNotEmpty ? _tasks.last.id : 0;
+    return _tasks.isNotEmpty ? _tasks.last!.id : 0;
   }
 
   ///getComplete is telling the function to get tasks that are complete or not
-  UnmodifiableListView<TaskModel> getSubtasksFor(
-    int id,
+  UnmodifiableListView<TaskModel?> getSubtasksFor(
+    int? id,
     // bool getComplete
   ) {
-    List<TaskModel> subtasks = [];
+    List<TaskModel?> subtasks = [];
 
     // bool tasksToChoose;
 
-    for (TaskModel task in _tasks) {
-      if (task.parent == id
+    for (TaskModel? task in _tasks) {
+      if (task!.parent == id
           // && task.complete == getComplete
           ) {
         subtasks.add(task);
@@ -47,7 +49,7 @@ class TaskListModel extends ChangeNotifier {
     return UnmodifiableListView(subtasks);
   }
 
-  TaskModel getTaskById(int id) {
+  TaskModel? getTaskById(int? id) {
     if (id == null) {
       return [
         {
@@ -64,7 +66,7 @@ class TaskListModel extends ChangeNotifier {
           .first;
     }
 
-    return _tasks.firstWhere((task) => task.id == id);
+    return _tasks.firstWhere((task) => task!.id == id);
   }
 
   addTask(Map<String, dynamic> taskData) async {
@@ -76,7 +78,7 @@ class TaskListModel extends ChangeNotifier {
     getAllTasks();
   }
 
-  Future<void> deleteTask(int id) async {
+  Future<void> deleteTask(int? id) async {
     // Delete this task from the db
     await _taskStore.record(id).delete(await _db);
 
@@ -86,10 +88,11 @@ class TaskListModel extends ChangeNotifier {
     return;
   }
 
-  Future<List<TaskModel>> getAllTasks() async {
-    final taskSnapshots = await _taskStore.find(await _db);
+  Future<List<TaskModel?>> getAllTasks() async {
+    final taskSnapshots = await (_taskStore.find(await _db));
+    // as FutureOr<List<RecordSnapshot<int, Map<String, Object>>>>);
 
-    List<TaskModel> tasks = taskSnapshots.map((snapshot) {
+    List<TaskModel?> tasks = taskSnapshots.map((snapshot) {
       final task = standardSerializers.deserializeWith(
           TaskModel.serializer, snapshot.value);
       return task;
@@ -103,9 +106,9 @@ class TaskListModel extends ChangeNotifier {
   }
 
   Future<void> toggleTaskComplete(TaskModel task) async {
-    task = task.rebuild((b) => b..complete = !b.complete);
-    Map<String, dynamic> taskData =
-        standardSerializers.serializeWith(TaskModel.serializer, task);
+    task = task.rebuild((b) => b..complete = !b.complete!);
+    Map<String, dynamic> taskData = standardSerializers.serializeWith(
+        TaskModel.serializer, task) as Map<String, dynamic>;
 
     await _taskStore.record(task.id).put(await _db, taskData);
     await getAllTasks();
